+++ src/types.ts
@@ -648,6 +648,8 @@ export interface LoadExchangeConfig {
   loadFromStore: boolean;
   TIFBufferSeconds: number;
   loadAssets?: Asset[];
+  doubleDownConnections?: Connection[];
+  blockhashCacheTimeoutSeconds?: number;
 }
 
 export function defaultLoadExchangeConfig(